+++ client/waatea2_client/lib/models/user_model.dart
@@ -6,13 +6,15 @@ class UserModel {
   final String email;
   final UserProfileModel profile;
   final int attendancePercentage;
+  final int caps;
 
   UserModel(
       {required this.pk,
       required this.name,
       required this.email,
       required this.profile,
-      required this.attendancePercentage});
+      required this.attendancePercentage,
+      required this.caps});
 
   factory UserModel.fromJson(Map<String, dynamic> json) {
     return UserModel(
@@ -20,7 +22,8 @@ class UserModel {
         name: json['name'],
         email: json['email'],
         profile: UserProfileModel.fromJson(json['profile']),
-        attendancePercentage: json['attendance_percentage']);
+        attendancePercentage: json['attendance_percentage'],
+        caps: json['caps']);
   }
 
   Map<String, dynamic> toJson() => {