+++ server/config/settings/base.py
@@ -279,6 +279,7 @@ SOCIALACCOUNT_FORMS = {"signup": "waatea_2.users.forms.UserSocialSignupForm"}
 REST_FRAMEWORK = {
     'DEFAULT_AUTHENTICATION_CLASSES': (
         'rest_framework.authentication.TokenAuthentication',
+'rest_framework.authentication.SessionAuthentication',
     ),
     'DEFAULT_PERMISSION_CLASSES': [
         'rest_framework.permissions.IsAuthenticated',